refactor(auth_service, teacher_service, security): split security logic into security.py
