+++ backend/app/services/auth_service.py
@@ -1,19 +1,10 @@
-from datetime import datetime, timedelta
-from jose import JWTError, jwt
-from passlib.context import CryptContext
 from sqlalchemy.orm import Session
 from fastapi import HTTPException, status
-from models import User 
-from schemas.auth import LoginRequest, UserCreate
+from models import User
+from schemas.auth import LoginRequest
+from datetime import timedelta
 from app.core.config import settings
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto") 
-
-def get_password_hash(password: str):
-    return pwd_context.hash(password)
-
-def verify_password(plain_password: str, hashed_password: str):
-    return pwd_context.verify(plain_password, hashed_password)
+from app.core.security import verify_password, create_access_token
 
 def get_user_by_username(db: Session, username: str):
     return db.query(User).filter(User.username == username).first()
@@ -24,13 +15,6 @@ def authenticate_user(db: Session, username: str, password: str):
         return None
     return user
 
-def create_access_token(data: dict, expires_delta: timedelta | None = None):
-    to_encode = data.copy()
-    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=15))
-    to_encode.update({"exp": expire})
-    return jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
-
-# Xử lý đăng nhập
 def login_service(db: Session, request: LoginRequest):
     user = authenticate_user(db, request.username, request.password)
     if not user:
